+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Display external ID in the log once a task has been sent
 - Content of column reported. Before was an icon and now it's just colored text.
+- Condense the space of the tables.
 
 ### Added