+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Display external ID in the log once a task has been sent
 - Content of column reported. Before was an icon and now it's just colored text.
 - Condense the space of the tables.
+- Minimize the size of the reported column.
 
 ### Added